+++ src/tensorEVD.h
@@ -11,10 +11,12 @@ SEXP get_dimnames(int nrow, int ncol,
                   int *icol1, int *icol2, int *icol,
                   SEXP dimnames1_, SEXP dimnames2_);
 
-void get_pos(int nA, int nB, int k, int *i, int *j);
+void get_pos(int nA, int nB, int k, int *i, int *j, int start);
 
-void get_kronecker_index(int nA, int nB, int *iA, int *iB, int ni, int *index);
+void get_kronecker_index(int nA, int nB, int *iA, int *iB, int ni, int *index, int start);
 
 void hadam_set(int n, double *a, double *dx, int *ix, double *dy, int *iy, double *dz);
 
 double dnorm_hadam_set(int n, double *dx, int *ix, double *dy, int *iy);
+
+void append_to_sorted_vector(int k, double *values, int *order);