+++ src/utils.c
@@ -1,5 +1,41 @@
 #include "tensorEVD.h"
 
+//==============================================================
+// Return sorted (descendant order) indices for a vector of
+// numeric values. The ordered position of the element k,
+// values[k], is found by comparing with the (ordered) values
+// of the previous k-1 values. This is, by finding the last
+// index order[j] (j=0,1,...,k) such that
+//                values[order[j]] > values[k]
+// where 'order' contains the ordered indices for the first
+//        k elements of 'values': values[0],...,values[k-1]
+// The new vector with sorted indices will contain the k+1
+// elements:
+//    order[0],...,order[j-1], k, order[j],...,order[k-1]
+//==============================================================
+void append_to_sorted_vector(int k, double *values, int *order)
+{
+  int j;
+
+  if(k==0){
+    order[0] = k;
+  }else{
+    j=0;
+    while(values[order[j]]>values[k]){
+      j++;
+      if(j == k){
+        break;
+      }
+    }
+    if(j < k){
+      memmove(order + j+1, order + j, (k-j)*sizeof(int));
+    }
+    order[j]=k;
+  }
+}
+
+//==============================================================
+//==============================================================
 SEXP get_dimnames(int nrow, int ncol,
                   int *irow1, int *irow2, int *irow,
                   int *icol1, int *icol2, int *icol,
@@ -66,15 +102,15 @@ SEXP get_dimnames(int nrow, int ncol,
 //     [1,2,...,nB,1,2,...,nB,...,1,2,...,nB] from B
 //
 //==============================================================
-void get_pos(int nA, int nB, int k, int *i, int *j)
+void get_pos(int nA, int nB, int k, int *i, int *j, int start)
 {
-  i[0] = floor(k/nB);
-  j[0] = k % nB;
+  i[0] = (int) floor(k/nB) + start;
+  j[0] = (k % nB) + start;
 }
 
 //==============================================================
 
-void get_kronecker_index(int nA, int nB, int *iA, int *iB, int ni, int *index)
+void get_kronecker_index(int nA, int nB, int *iA, int *iB, int ni, int *index, int start)
 {
   int i, j;
 
@@ -82,16 +118,99 @@ void get_kronecker_index(int nA, int nB, int *iA, int *iB, int ni, int *index)
     int tmp = 0;
     for(i=0; i<nA; i++){
       for(j=0; j<nB; j++){
-        iA[tmp] = i;
-        iB[tmp] = j;
+        iA[tmp] = i + start;
+        iB[tmp] = j + start;
         tmp++;
       }
     }
   }else{
     for(i=0; i<ni; i++){
-      get_pos(nA, nB, index[i], iA+i, iB+i);
+      get_pos(nA, nB, index[i], iA+i, iB+i, start);
     }
   }
+}
+
+SEXP R_kronecker_index(SEXP nrowA_, SEXP ncolA_,
+                       SEXP nrowB_, SEXP ncolB_,
+                       SEXP rows_, SEXP cols_,
+                       SEXP swap_, SEXP zerobased_)
+{
+  int nprotect = 6;
+
+  int nrowA = INTEGER_VALUE(nrowA_);
+  int ncolA = INTEGER_VALUE(ncolA_);
+  int nrowB = INTEGER_VALUE(nrowB_);
+  int ncolB = INTEGER_VALUE(ncolB_);
+  int swap = asLogical(swap_);
+
+  int start = asLogical(zerobased_) ? 0 : 1;
+  int nirow = Rf_length(rows_);
+  int nicol = Rf_length(cols_);
+
+  PROTECT(rows_ = AS_INTEGER(rows_));
+  int *irow = INTEGER_POINTER(rows_);
+
+  PROTECT(cols_ = AS_INTEGER(cols_));
+  int *icol = INTEGER_POINTER(cols_);
+
+  int nrow = (nirow == 0) ? nrowA*nrowB : nirow;
+  int ncol = (nicol == 0) ? ncolA*ncolB : nicol;
+
+  SEXP irowA_, irowB_, icolA_, icolB_;
+  int *irowA, *irowB, *icolA, *icolB;
+
+  // Get positions for rows: irowA and irowB
+  irowA_ = PROTECT(Rf_allocVector(INTSXP, nrow));
+  irowA = INTEGER_POINTER(irowA_);
+
+  irowB_ = PROTECT(Rf_allocVector(INTSXP, nrow));
+  irowB = INTEGER_POINTER(irowB_);
+
+  // Rprintf(" Creating indices for rows: irowA and irowB ...\n");
+  if(swap){   // kronecker(B,A)
+    get_kronecker_index(nrowB, nrowA, irowB, irowA, nirow, irow, start);
+  }else{      // kronecker(A,B)
+    get_kronecker_index(nrowA, nrowB, irowA, irowB, nirow, irow, start);
+  }
+
+  // Get positions for cols: icolA and icolB
+  if((nrowA==ncolA) && (nrowB==ncolB) && ((nirow+nicol)==0)){
+    // Rprintf(" Recycling indices for cols because both are squared matrices ...\n");
+    icolA_ = irowA_;
+    icolB_ = irowB_;
+  }else{
+    // Rprintf(" Creating indices for cols: icolA and icolB ...\n");
+    icolA_ = PROTECT(Rf_allocVector(INTSXP, ncol));
+    icolA = INTEGER_POINTER(icolA_);
+
+    icolB_ = PROTECT(Rf_allocVector(INTSXP, ncol));
+    icolB = INTEGER_POINTER(icolB_);
+    nprotect += 2;
+
+    if(swap){   // kronecker(B,A)
+      get_kronecker_index(ncolB, ncolA, icolB, icolA, nicol, icol, start);
+    }else{      // kronecker(A,B)
+      get_kronecker_index(ncolA, ncolB, icolA, icolB, nicol, icol, start);
+    }
+  }
+
+  // Rprintf(" Outputting irowA, icolA, irowB, and icolB ...\n");
+  SEXP list_ = PROTECT(Rf_allocVector(VECSXP, 4));
+  SET_VECTOR_ELT(list_, 0, irowA_);
+  SET_VECTOR_ELT(list_, 1, icolA_);
+  SET_VECTOR_ELT(list_, 2, irowB_);
+  SET_VECTOR_ELT(list_, 3, icolB_);
+
+  SEXP names_ = PROTECT(Rf_allocVector(VECSXP, 4));
+  SET_VECTOR_ELT(names_, 0, mkChar("irowA"));
+  SET_VECTOR_ELT(names_, 1, mkChar("icolA"));
+  SET_VECTOR_ELT(names_, 2, mkChar("irowB"));
+  SET_VECTOR_ELT(names_, 3, mkChar("icolB"));
+  setAttrib(list_, R_NamesSymbol, names_);
+
+  UNPROTECT(nprotect);
+
+  return(list_);
 }
 
 //====================================================================